Add base test for extending code highlighting
