+++ lib/template.js
@@ -339,7 +339,6 @@ TemplateEngine.prototype.addBlock = function(name, block) {
 
 // Add multiple blocks
 TemplateEngine.prototype.addBlocks = function(blocks) {
-    console.log('add blocks', blocks);
     _.each(blocks, function(block, name) {
         this.addBlock(name, block);
     }, this);
@@ -351,7 +350,6 @@ TemplateEngine.prototype.applyBlock = function(name, blk) {
     var func, block, func, r;
 
     block = this.blocks[name];
-    console.log('applyBlock', name, block);
     if (!block) throw new Error('Block not found "'+name+'"');
     if (_.isString(blk)) {
         blk = {
