+++ test/books/highlight/README.md
@@ -1,3 +1,7 @@
 # Readme
 
 Default description for the book.
+
+```
+test
+```
